+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 ## Under development
 * TBD
 
+# 1.0.2 - 25 August 2016
+* Allow free form tasks to be supplied without a block
+* Ensure non-free form tasks are called with a block
+* Add separate clean/compile rake Tasks
+* Rename existing Rake task to execute
+* Try and avoid DSL conflicts
+
 ## 1.0.1 - 24 August 2016
 * Bug fix/missing dependency